Refactor budget list view in smaller components
